pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Multi-agent deep research system using Claude and Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "google-genai>=1.0.0",
]
